+++ index.js
@@ -3,93 +3,35 @@
 // This is dangerous as it hides accidentally undefined variables.
 // We blacklist the globals that we deem potentially confusing.
 // To use them, explicitly reference them, e.g. `window.name` or `window.status`.
-var restrictedGlobals = [
-  'addEventListener',
-  'blur',
-  'close',
-  'closed',
-  'confirm',
-  'defaultStatus',
-  'defaultstatus',
-  'event',
-  'external',
-  'find',
-  'focus',
-  'frameElement',
-  'frames',
-  'history',
-  'innerHeight',
-  'innerWidth',
-  'length',
-  'location',
-  'locationbar',
-  'menubar',
-  'moveBy',
-  'moveTo',
-  'name',
-  'onblur',
-  'onerror',
-  'onfocus',
-  'onload',
-  'onresize',
-  'onunload',
-  'open',
-  'opener',
-  'opera',
-  'outerHeight',
-  'outerWidth',
-  'pageXOffset',
-  'pageYOffset',
-  'parent',
-  'print',
-  'removeEventListener',
-  'resizeBy',
-  'resizeTo',
-  'screen',
-  'screenLeft',
-  'screenTop',
-  'screenX',
-  'screenY',
-  'scroll',
-  'scrollbars',
-  'scrollBy',
-  'scrollTo',
-  'scrollX',
-  'scrollY',
-  'self',
-  'status',
-  'statusbar',
-  'stop',
-  'toolbar',
-  'top',
-];
+var restrictedGlobals = require('confusing-browser-globals');
 
 module.exports = {
   root: true,
 
-  parser: '@typescript-eslint/parser',
+  parser: 'babel-eslint',
 
-  parserOptions: {},
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
 
-  extends: [
-    'airbnb',
-    'plugin:@typescript-eslint/recommended',
-    'prettier',
-    'prettier/react',
-    'prettier/@typescript-eslint',
-  ],
+  extends: ['airbnb', 'prettier', 'prettier/react'],
 
   plugins: [
     'import',
+    'flowtype',
     'jsx-a11y',
     'react',
     'react-hooks',
     'prettier',
-    '@typescript-eslint',
   ],
 
   env: {
     browser: true,
+    commonjs: true,
     es6: true,
     jest: true,
     node: true,
@@ -103,8 +45,11 @@ module.exports = {
     'import/extensions': ['.js', '.jsx'],
     'import/resolver': {
       node: {
-        extensions: ['.js', '.jsx', '.json', '.ts', '.tsx'],
+        extensions: ['.js', '.jsx', '.json'],
       },
+    },
+    react: {
+      version: 'detect',
     },
   },
 
@@ -113,15 +58,6 @@ module.exports = {
     'no-restricted-globals': ['error'].concat(restrictedGlobals),
     'no-underscore-dangle': 'off',
     'prettier/prettier': 'error',
-    // 'prettier/prettier': ['error', {
-    //   trailingComa: 'es5',
-    //   tabWidth: 2,
-    //   semi: true,
-    //   singleQuote: true,
-    //   jsxBracketSameLine: false,
-    //   bracketSpacing: true,
-    //   arrowParens: 'always'
-    // }],
     'react/destructuring-assignment': [
       'error',
       'always',
@@ -154,23 +90,66 @@ module.exports = {
 
   overrides: [
     {
-      files: ['*.js', '*.jsx'],
-      rules: {
-        '@typescript-eslint/no-var-requires': 'off',
+      files: ['**/*.ts', '**/*.tsx'],
+      parser: '@typescript-eslint/parser',
+      parserOptions: {
+        ecmaVersion: 2018,
+        sourceType: 'module',
+        emaFeatures: {
+          jsx: true,
+        },
+
+        // typescript-eslint specific options
+        warnOnUnsupportedTypeScriptVersion: true,
       },
-    },
-    {
-      files: ['*.ts', '*.tsx'],
+      plugins: ['@typescript-eslint'],
       rules: {
+        'import/no-unresolved': 'off',
+        'react/prop-types': 'off',
+
+        // Default Recommeneded TS Rules
+        // See: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/src/configs/recommended.json
+
+        '@typescript-eslint/adjacent-overload-signatures': 'error',
+        '@typescript-eslint/array-type': 'error',
+        '@typescript-eslint/ban-types': 'error',
+        camelcase: 'off',
+        '@typescript-eslint/camelcase': 'error',
+        '@typescript-eslint/class-name-casing': 'error',
+        // '@typescript-eslint/explicit-function-return-type': 'warn',
+        // '@typescript-eslint/explicit-member-accessibility': 'error',
+        indent: 'off',
+        // '@typescript-eslint/indent': 'error',
+        '@typescript-eslint/interface-name-prefix': 'error',
+        // '@typescript-eslint/member-delimiter-style': 'error',
+        '@typescript-eslint/no-angle-bracket-type-assertion': 'error',
+        'no-array-constructor': 'off',
+        '@typescript-eslint/no-array-constructor': 'error',
+        '@typescript-eslint/no-empty-interface': 'error',
+        '@typescript-eslint/no-explicit-any': 'warn',
+        '@typescript-eslint/no-inferrable-types': 'error',
+        '@typescript-eslint/no-misused-new': 'error',
+        '@typescript-eslint/no-namespace': 'error',
+        '@typescript-eslint/no-non-null-assertion': 'error',
+        '@typescript-eslint/no-object-literal-type-assertion': 'error',
+        '@typescript-eslint/no-parameter-properties': 'error',
+        '@typescript-eslint/no-triple-slash-reference': 'error',
+        'no-unused-vars': 'off',
+        '@typescript-eslint/no-unused-vars': 'warn',
+        '@typescript-eslint/no-use-before-define': 'error',
+        '@typescript-eslint/no-var-requires': 'error',
+        '@typescript-eslint/prefer-interface': 'error',
+        '@typescript-eslint/prefer-namespace-keyword': 'error',
+        // '@typescript-eslint/type-annotation-spacing': 'error',
+
+        // Prettier TS Disabled Rules
+        '@typescript-eslint/indent': 'off',
+        '@typescript-eslint/member-delimiter-style': 'off',
+        '@typescript-eslint/type-annotation-spacing': 'off',
+
+        // Smartcare Sanity TS Disabled Rules
         '@typescript-eslint/explicit-function-return-type': 'off',
         '@typescript-eslint/explicit-member-accessibility': 'off',
-        'import/no-unresolved': 'off',
-      },
-    },
-    {
-      files: ['*.tsx'],
-      rules: {
-        'react/prop-types': 'off',
       },
     },
   ],