+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+<a name="5.1.0"></a>
+
+# [5.0.1](https://github.com/smart-care/eslint-config-smartcare/compare/v5.0.0...v5.1.0) (2019-04-04)
+
+### Features
+
+- Replaced our custom global restriction list with `confusing-browser-globals`.
+
+### Fixes
+
+- Added back Flow support.
+- Make TypeScript linting only apply to TypeScript files.
+
 <a name="5.0.0"></a>
 
 # [5.0.0](https://github.com/smart-care/eslint-config-smartcare/compare/v4.0.0...v5.0.0) (2019-04-03)
@@ -5,5 +18,3 @@
 ### BREAKING CHANGES
 
 Version 5 represents our move over to TypeScript for all future projects at Smartcare. So this versions `peerDependencies` have changed. Please refer to the readme to check which packages are needed.
-
-For our config without TypeScript which is no longer supported going forward, see the [v4 branch](https://github.com/smart-care/eslint-config-smartcare/tree/v4).